Reorganized to help with readability
